+++ src/main/java/pairmatching/Application.java
@@ -1,5 +1,7 @@
 package pairmatching;
 
+import pairmatching.domain.PairMatchingController;
+
 public class Application {
     public static void main(String[] args) {
         PairMatchingController.run();
