--- src/main/java/pairmatching/Crew.java
@@ -1,6 +1,0 @@
-package pairmatching;
-
-public class Crew {
-    private Course course;
-    private String name;
-}
