+++ src/main/java/pairmatching/domain/CrewRepository.java
@@ -1,10 +1,11 @@
-package pairmatching;
+package pairmatching.domain;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import pairmatching.domain.option.Course;
 
 public class CrewRepository {
     private List<String> backend = new ArrayList<>();
