+++ src/test/java/pairmatching/PairMatchingServiceTest.java
@@ -1,9 +1,9 @@
 package pairmatching;
 
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import pairmatching.domain.PairMatchingService;
 
 class PairMatchingServiceTest {
